requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
prompt-toolkit>=3.0.0
rich>=13.0.0
pyyaml>=6.0
pydantic>=2.0.0
click>=8.0.0
